requirements.txt
cryptography==36.0.2
fastapi==0.109.2
Flask==2.3.2
matplotlib==3.7.2
numpy==1.24.3
pandas==2.2.0
Pillow==9.0.1
Pillow==10.2.0
pycryptodome==3.19.1
pymongo==4.6.1
pytesseract==0.3.10
sympy==1.12
tensorflow==2.13.0
tensorflow_intel==2.13.0
uvicorn==0.27.1
waitress==2.1.2
